conda_requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
altgraph=0.17.3=py310haa95532_0
blas=1.0=mkl
brotli=1.0.9=h2bbff1b_7
brotli-bin=1.0.9=h2bbff1b_7
bzip2=1.0.8=he774522_0
ca-certificates=2023.01.10=haa95532_0
certifi=2022.12.7=py310haa95532_0
contourpy=1.0.5=py310h59b6b97_0
cycler=0.11.0=pyhd3eb1b0_0
eigen=3.3.7=h59b6b97_1
fonttools=4.25.0=pyhd3eb1b0_0
freetype=2.12.1=ha860e81_0
future=0.18.2=py310haa95532_1
glib=2.69.1=h5dc1a3c_2
gst-plugins-base=1.18.5=h9e645db_0
gstreamer=1.18.5=hd78058f_0
hdf5=1.12.1=h1756f20_2
icc_rt=2022.1.0=h6049295_2
icu=58.2=ha925a31_3
intel-openmp=2021.4.0=haa95532_3556
jpeg=9e=h2bbff1b_0
kiwisolver=1.4.4=py310hd77b12b_0
lerc=3.0=hd77b12b_0
libbrotlicommon=1.0.9=h2bbff1b_7
libbrotlidec=1.0.9=h2bbff1b_7
libbrotlienc=1.0.9=h2bbff1b_7
libclang=12.0.0=default_h627e005_2
libdeflate=1.8=h2bbff1b_5
libffi=3.4.2=hd77b12b_6
libiconv=1.16=h2bbff1b_2
libogg=1.3.5=h2bbff1b_1
libpng=1.6.37=h2a8f88b_0
libprotobuf=3.20.3=h23ce68f_0
libtiff=4.5.0=h6c2663c_1
libvorbis=1.3.7=he774522_0
libwebp=1.2.4=h2bbff1b_0
libwebp-base=1.2.4=h2bbff1b_0
libxml2=2.9.14=h0ad7f3c_0
libxslt=1.1.35=h2bbff1b_0
lz4-c=1.9.4=h2bbff1b_0
matplotlib=3.6.2=py310haa95532_0
matplotlib-base=3.6.2=py310h1094b8e_0
mkl=2021.4.0=haa95532_640
mkl-service=2.4.0=py310h2bbff1b_0
mkl_fft=1.3.1=py310ha0764ea_0
mkl_random=1.2.2=py310h4ed8f06_0
munkres=1.1.4=py_0
numpy=1.23.5=py310h60c9a35_0
numpy-base=1.23.5=py310h04254f7_0
opencv=4.6.0=py310ha7641e4_2
openssl=1.1.1s=h2bbff1b_0
packaging=22.0=py310haa95532_0
pcre=8.45=hd77b12b_0
pefile=2022.5.30=py310haa95532_0
pillow=9.3.0=py310hd77b12b_2
pip=22.3.1=py310haa95532_0
ply=3.11=py310haa95532_0
pyinstaller=5.6.2=py310h2bbff1b_0
pyinstaller-hooks-contrib=2022.14=py310haa95532_0
pyocr=0.8.3=pypi_0
pyparsing=3.0.9=py310haa95532_0
pyqt=5.15.7=py310hd77b12b_0
pyqt5-sip=12.11.0=py310hd77b12b_0
python=3.10.9=h966fe2a_0
python-dateutil=2.8.2=pyhd3eb1b0_0
pywin32=305=py310h2bbff1b_0
pywin32-ctypes=0.2.0=py310haa95532_1000
qt-main=5.15.2=he8e5bd7_7
qt-webengine=5.15.9=hb9a9bb5_5
qtwebkit=5.212=h3ad3cdb_4
setuptools=65.6.3=py310haa95532_0
sip=6.6.2=py310hd77b12b_0
six=1.16.0=pyhd3eb1b0_1
sqlite=3.40.1=h2bbff1b_0
tk=8.6.12=h2bbff1b_0
toml=0.10.2=pyhd3eb1b0_0
tornado=6.2=py310h2bbff1b_0
tzdata=2022g=h04d1e81_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
wheel=0.37.1=pyhd3eb1b0_0
wincertstore=0.2=py310haa95532_2
xz=5.2.10=h8cc25b3_1
zlib=1.2.13=h8cc25b3_0
zstd=1.5.2=h19a0ad4_0
